Fix failure to execute a `program' when a directory `program' exists
If a directory named is `program' exists in a current directory,
`program' cannot be executed because the directory `program' is
searched first before the executable `program'.

This occurs when executing `git add -p' in a directory in which a
directory `git' exists.

This is a bug of kLIBC v0.6.6 which is already fixed, but the fixed
kLIBC has not been released, yet.

    modified:   os2/os2.c